# services/notification/api/openapi.yaml
# DO NOT EDIT — generated by tools/specgen.
# Source: services/notification/internal/adapter/inbound/http/endpoints.go
openapi: 3.1.0
info:
  title: notification
  version: 1.0.0
paths:
  /v1/notifications:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsListResponse'
          description: success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: List notifications for a recipient (?recipient=<email>)
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchNotificationRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
          description: success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
      summary: Dispatch a notification (called by Argo Events sensors)
  /v1/notifications/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
          description: success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
      summary: Get a single notification by ID
components:
  schemas:
    DispatchNotificationRequest:
      type: object
      properties:
        body:
          type: string
        channel:
          type: string
        idempotency_key:
          type: string
        recipient:
          type: string
        subject:
          type: string
      required:
        - recipient
        - channel
        - subject
        - body
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    NotificationResponse:
      type: object
      properties:
        body:
          type: string
        channel:
          type: string
        id:
          type: string
        recipient:
          type: string
        sent_at:
          type: object
        status:
          type: string
        subject:
          type: string
      required:
        - id
        - recipient
        - channel
        - subject
        - body
        - status
    NotificationsListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            type: object
            properties:
              body:
                type: string
              channel:
                type: string
              id:
                type: string
              recipient:
                type: string
              sent_at:
                type: object
              status:
                type: string
              subject:
                type: string
            required:
              - id
              - recipient
              - channel
              - subject
              - body
              - status
      required:
        - notifications